refacotr(vuex-class): convert action code to generic method factory
